+++ lib/domain_layer/errors.dart
@@ -1,0 +1,3 @@
+library errors;
+
+export 'errors/biometrics_not_available_exception.dart';