+++ lib/presentation_layer/app.dart
@@ -1,10 +1,10 @@
 library app;
 
 export './app/app_builder.dart';
-export './app/app_configuration.dart';
-export './app/app_localization_configuration.dart';
-export './app/app_navigation_configuration.dart';
-export './app/app_ssl_configuration.dart';
-export './app/app_theme_configuration.dart';
 export './app/bank_app.dart';
+export './app/configuration/app_localization_configuration.dart';
+export './app/configuration/app_navigation_configuration.dart';
+export './app/configuration/app_theme_configuration.dart';
 export './app/ssl_configuration_bank_app_wrapper.dart';
+export 'app/configuration/app_configuration.dart';
+export 'app/configuration/app_ssl_configuration.dart';