+++ lib/presentation_layer/app/configuration/authentication_method_configuration.dart
@@ -1,0 +1,5 @@
+/// The supported authentication methods.
+enum AuthenticationMethodConfiguration {
+  /// The app uses access pin authentication.
+  accessPin;
+}