+++ lib/presentation_layer/creators/storage/storage_creator.dart
@@ -12,11 +12,9 @@ class StorageCreator implements CubitCreator {
   final SaveAuthenticationSettingUseCase _saveAuthenticationSettingUseCase;
   final LoadOcraSecretKeyUseCase _loadOcraSecretKeyUseCase;
   final SaveOcraSecretKeyUseCase _saveOcraSecretKeyUseCase;
-  final LoadAccessPinUseCase _loadAccessPinUseCase;
-  final SaveAccessPinUseCase _saveAccessPinUseCase;
+  final SaveAccessPinForBiometricsUseCase _saveAccessPinUseCase;
   final SetBrightnessUseCase _setBrightnessUseCase;
   final LoadBrightnessUseCase _loadBrightnessUseCase;
-  final ToggleBiometricsUseCase _toggleBiometricsUseCase;
   final LoadLoyaltyTutorialCompletionUseCase
       _loadLoyaltyTutorialCompletionUseCase;
   final SetLoyaltyTutorialCompletionUseCase
@@ -33,11 +31,9 @@ class StorageCreator implements CubitCreator {
     required SaveAuthenticationSettingUseCase saveAuthenticationSettingUseCase,
     required LoadOcraSecretKeyUseCase loadOcraSecretKeyUseCase,
     required SaveOcraSecretKeyUseCase saveOcraSecretKeyUseCase,
-    required LoadAccessPinUseCase loadAccessPinUseCase,
-    required SaveAccessPinUseCase saveAccessPinUseCase,
+    required SaveAccessPinForBiometricsUseCase saveAccessPinUseCase,
     required SetBrightnessUseCase setBrightnessUseCase,
     required LoadBrightnessUseCase loadBrightnessUseCase,
-    required ToggleBiometricsUseCase toggleBiometricsUseCase,
     required LoadLoyaltyTutorialCompletionUseCase
         loadLoyaltyTutorialCompletionUseCase,
     required SetLoyaltyTutorialCompletionUseCase
@@ -50,11 +46,9 @@ class StorageCreator implements CubitCreator {
         _saveAuthenticationSettingUseCase = saveAuthenticationSettingUseCase,
         _loadOcraSecretKeyUseCase = loadOcraSecretKeyUseCase,
         _saveOcraSecretKeyUseCase = saveOcraSecretKeyUseCase,
-        _loadAccessPinUseCase = loadAccessPinUseCase,
         _saveAccessPinUseCase = saveAccessPinUseCase,
         _setBrightnessUseCase = setBrightnessUseCase,
         _loadBrightnessUseCase = loadBrightnessUseCase,
-        _toggleBiometricsUseCase = toggleBiometricsUseCase,
         _loadLoyaltyTutorialCompletionUseCase =
             loadLoyaltyTutorialCompletionUseCase,
         _setLoyaltyTutorialCompletionUseCase =
@@ -70,11 +64,9 @@ class StorageCreator implements CubitCreator {
         saveAuthenticationSettingUseCase: _saveAuthenticationSettingUseCase,
         loadOcraSecretKeyUseCase: _loadOcraSecretKeyUseCase,
         saveOcraSecretKeyUseCase: _saveOcraSecretKeyUseCase,
-        loadAccessPinUseCase: _loadAccessPinUseCase,
-        saveAccessPinUseCase: _saveAccessPinUseCase,
+        saveAccessPinForBiometricsUseCase: _saveAccessPinUseCase,
         setBrightnessUseCase: _setBrightnessUseCase,
         loadBrightnessUseCase: _loadBrightnessUseCase,
-        toggleBiometricsUseCase: _toggleBiometricsUseCase,
         loadLoyaltyTutorialCompletionUseCase:
             _loadLoyaltyTutorialCompletionUseCase,
         setLoyaltyTutorialCompletionUseCase: