+++ test/cubits/storage/storage_cubit_test.dart
@@ -25,8 +25,6 @@ class MockLoadLastLoggedUserUseCase extends Mock
 class MockLoadOcraSecretKeyUseCase extends Mock
     implements LoadOcraSecretKeyUseCase {}
 
-class MockLoadAccessPinUseCase extends Mock implements LoadAccessPinUseCase {}
-
 class MockRemoveUserUseCase extends Mock implements RemoveUserUseCase {}
 
 class MockSaveAuthenticationSettingUseCase extends Mock
@@ -35,14 +33,12 @@ class MockSaveAuthenticationSettingUseCase extends Mock
 class MockSaveOcraSecretKeyUseCase extends Mock
     implements SaveOcraSecretKeyUseCase {}
 
-class MockSaveAccessPinUseCase extends Mock implements SaveAccessPinUseCase {}
+class MockSaveAccessPinUseCase extends Mock
+    implements SaveAccessPinForBiometricsUseCase {}
 
 class MockSaveUserUseCase extends Mock implements SaveUserUseCase {}
 
 class MockSetBrightnessUseCase extends Mock implements SetBrightnessUseCase {}
-
-class MockToggleBiometricsUseCase extends Mock
-    implements ToggleBiometricsUseCase {}
 
 class MockLoadLoyaltyTutorialCompletionUseCase extends Mock
     implements LoadLoyaltyTutorialCompletionUseCase {}
@@ -56,7 +52,6 @@ final _loadBrightnessUseCase = MockLoadBrightnessUseCase();
 final _loadLoggedInUsersUseCase = MockLoadLoggedInUsersUseCase();
 final _loadLastLoggedUserUseCase = MockLoadLastLoggedUserUseCase();
 final _loadOcraSecretKeyUseCase = MockLoadOcraSecretKeyUseCase();
-final _loadAccessPinUseCase = MockLoadAccessPinUseCase();
 final _removeUserUseCase = MockRemoveUserUseCase();
 final _saveAuthenticationSettingUseCase =
     MockSaveAuthenticationSettingUseCase();
@@ -64,7 +59,6 @@ final _saveOcraSecretKeyUseCase = MockSaveOcraSecretKeyUseCase();
 final _saveAccessPinUseCase = MockSaveAccessPinUseCase();
 final _saveUserUseCase = MockSaveUserUseCase();
 final _setBrightnessUseCase = MockSetBrightnessUseCase();
-final _toggleBiometricsUseCase = MockToggleBiometricsUseCase();
 final _loadLoyaltyTutorialCompletionUseCase =
     MockLoadLoyaltyTutorialCompletionUseCase();
 final _setLoyaltyTutorialCompletionUseCase =
@@ -116,14 +110,12 @@ StorageCubit createStorageCubit() => StorageCubit(
       loadAuthenticationSettingsUseCase: _loadAuthenticationSettingsUseCase,
       loadBrightnessUseCase: _loadBrightnessUseCase,
       loadOcraSecretKeyUseCase: _loadOcraSecretKeyUseCase,
-      loadAccessPinUseCase: _loadAccessPinUseCase,
       removeUserUseCase: _removeUserUseCase,
       saveAuthenticationSettingUseCase: _saveAuthenticationSettingUseCase,
       saveOcraSecretKeyUseCase: _saveOcraSecretKeyUseCase,
-      saveAccessPinUseCase: _saveAccessPinUseCase,
+      saveAccessPinForBiometricsUseCase: _saveAccessPinUseCase,
       saveUserUseCase: _saveUserUseCase,
       setBrightnessUseCase: _setBrightnessUseCase,
-      toggleBiometricsUseCase: _toggleBiometricsUseCase,
       loadLoyaltyTutorialCompletionUseCase:
           _loadLoyaltyTutorialCompletionUseCase,
       setLoyaltyTutorialCompletionUseCase: _setLoyaltyTutorialCompletionUseCase,
@@ -145,7 +137,6 @@ void main() {
   group('Biometrics', _biometricsTests);
   group('Application Settings', _applicationSettingsTests);
   group('OCRA key', _ocraKeyTests);
-  group('Access pin', _accessPinTests);
   group('Loyalty tutorial completion', _loyaltyCompletionTests);
 }
 
@@ -320,14 +311,12 @@ void _saveDataTests() {
       loadAuthenticationSettingsUseCase: _loadAuthenticationSettingsUseCase,
       loadBrightnessUseCase: _loadBrightnessUseCase,
       loadOcraSecretKeyUseCase: _loadOcraSecretKeyUseCase,
-      loadAccessPinUseCase: _loadAccessPinUseCase,
       removeUserUseCase: _removeUserUseCase,
       saveAuthenticationSettingUseCase: _saveAuthenticationSettingUseCase,
       saveOcraSecretKeyUseCase: _saveOcraSecretKeyUseCase,
-      saveAccessPinUseCase: _saveAccessPinUseCase,
+      saveAccessPinForBiometricsUseCase: _saveAccessPinUseCase,
       saveUserUseCase: _saveUserUseCase,
       setBrightnessUseCase: _setBrightnessUseCase,
-      toggleBiometricsUseCase: _toggleBiometricsUseCase,
       loadLoyaltyTutorialCompletionUseCase:
           _loadLoyaltyTutorialCompletionUseCase,
       setLoyaltyTutorialCompletionUseCase: _setLoyaltyTutorialCompletionUseCase,
@@ -467,15 +456,15 @@ void _saveDataTests() {
 void _biometricsTests() {
   setUp(() {
     when(
-      () => _toggleBiometricsUseCase(
-        isBiometricsActive: _useBiometrics,
+      () => _saveAuthenticationSettingUseCase(
+        useBiometrics: _useBiometrics,
       ),
     ).thenAnswer(
         (_) async => AuthenticationSettings(useBiometrics: _useBiometrics));
 
     when(
-      () => _toggleBiometricsUseCase(
-        isBiometricsActive: _failBiometrics,
+      () => _saveAuthenticationSettingUseCase(
+        useBiometrics: _failBiometrics,
       ),
     ).thenThrow(Exception('Some error'));
   });
@@ -504,8 +493,8 @@ void _biometricsTests() {
     ],
     verify: (c) {
       verify(
-        () => _toggleBiometricsUseCase(
-          isBiometricsActive: _useBiometrics,
+        () => _saveAuthenticationSettingUseCase(
+          useBiometrics: _useBiometrics,
         ),
       ).called(1);
     },
@@ -534,8 +523,8 @@ void _biometricsTests() {
     ],
     verify: (c) {
       verify(
-        () => _toggleBiometricsUseCase(
-          isBiometricsActive: _failBiometrics,
+        () => _saveAuthenticationSettingUseCase(
+          useBiometrics: _failBiometrics,
         ),
       ).called(1);
     },
@@ -769,91 +758,6 @@ _ocraKeyTests() {
       verify(
         () => _saveOcraSecretKeyUseCase(
           value: failedOcraKey,
-        ),
-      ).called(1);
-    },
-  ); // should handle exceptions
-}
-
-_accessPinTests() {
-  final successfulAccessPin = 'successfulAccessPin';
-  final failedAccessPin = 'failedAccessPin';
-
-  setUp(() {
-    when(
-      _loadAccessPinUseCase,
-    ).thenAnswer((_) async => successfulAccessPin);
-
-    when(
-      () => _saveAccessPinUseCase(
-        value: successfulAccessPin,
-      ),
-    ).thenAnswer((_) async => true);
-
-    when(
-      () => _saveAccessPinUseCase(
-        value: failedAccessPin,
-      ),
-    ).thenThrow(Exception());
-  });
-
-  blocTest<StorageCubit, StorageState>(
-    'should load access pin',
-    build: createStorageCubit,
-    act: (c) => c.loadAccessPin(),
-    expect: () => [
-      StorageState(
-        busy: true,
-      ),
-      StorageState(
-        accessPin: successfulAccessPin,
-      ),
-    ],
-    verify: (c) {
-      verify(
-        _loadAccessPinUseCase,
-      ).called(1);
-    },
-  ); // should load ocra key
-
-  blocTest<StorageCubit, StorageState>(
-    'should save access pin',
-    build: createStorageCubit,
-    act: (c) => c.saveAccessPin(successfulAccessPin),
-    expect: () => [
-      StorageState(
-        busy: true,
-      ),
-      StorageState(
-        accessPin: successfulAccessPin,
-      ),
-    ],
-    verify: (c) {
-      verify(
-        () => _saveAccessPinUseCase(
-          value: successfulAccessPin,
-        ),
-      ).called(1);
-    },
-  ); // should save ocra key
-
-  blocTest<StorageCubit, StorageState>(
-    'should handle exceptions when saving access pin',
-    build: createStorageCubit,
-    act: (c) => c.saveAccessPin(failedAccessPin),
-    expect: () => [
-      StorageState(
-        busy: true,
-      ),
-      StorageState(),
-    ],
-    errors: () => [
-      isA<Exception>(),
-    ],
-    verify: (c) {
-      verify(
-        () => _saveAccessPinUseCase(
-          value: failedAccessPin,
         ),
       ).called(1);
     },